Restructured logger configuration to config directory.
