+++ src/index.ts
@@ -4,7 +4,7 @@ dotenv.config();
 
 import mongoose from "mongoose";
 
-import logger from "./logger";
+import logger from "./config/logger";
 
 import { app } from "./app";
